refactor(mjh): one component per file + repo-owned persistence (PR-gate compliance)
Extract NoSessionView / SessionLoadErrorView / SessionLoadingView /
ActiveSessionView / ResumeRefinementHeader out of the ResumeRefinement
page and ResumeJobOption out of SessionStartPanel; move the remaining
direct db.flush/commit calls in session_helpers and session_turn_service
into session_repo primitives (clear_pending_state, bump_turn_count,
persist_prefetch_results).
